==> Figure5/01_NMF_CD4_T_cell_Projection/02_Extract_NMF_Scores.ipynb <==
suppressPackageStartupMessages(library(plyr))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
library(parallel)

# BR1

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V2/Dataset/scRNA_meta_data-2024-05-09.csv")
meta_data_subset<-meta_data %>% filter(Covid_exclusion=='no',
                                       sample.visitName%in% c('Flu Year 1 Day 0'))

df_list<-mclapply(meta_data_subset$pbmc_sample_id,function(x){
df<-read.csv(paste0("NMF_Score_BRI/",x,".csv"))
df$pbmc_sample_id=x
return(df)
},mc.cores=60)

df_combined<-do.call(rbind,df_list)
df_combined_longformat<- pivot_longer(df_combined, cols = starts_with("NMF"), names_to = "NMF_Scores", values_to = "value")

### Get Mean NMF scores

df_combined_longformat<-df_combined_longformat %>% group_by(pbmc_sample_id,NMF_Scores,AIFI_L3) %>% summarise(AverageScore=mean(value))

write.csv(df_combined_longformat,'NMF_Scores_Y1D0_BRI.csv')

# SF4

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V2/Dataset/SF4_meta_data-2024-05-05.csv")
meta_data$ageAtEnrollment[meta_data$ageAtEnrollment == '89+'] <- '90'
meta_data$ageAtEnrollment<-as.numeric(meta_data$ageAtEnrollment)
meta_data<-meta_data[!is.na(meta_data$ageAtEnrollment),]
meta_data$Ages<-meta_data$ageAtEnrollment

df_list<-mclapply(meta_data$pbmc_sample_id,function(x){
df<-read.csv(paste0("NMF_Score_SF4/",x,".csv"))
df$pbmc_sample_id=x
return(df)
},mc.cores=60)

df_combined<-do.call(rbind,df_list)
df_combined_longformat<- pivot_longer(df_combined, cols = starts_with("NMF"), names_to = "NMF_Scores", values_to = "value")

### Get Mean NMF scores

df_combined_longformat<-df_combined_longformat %>% group_by(pbmc_sample_id,NMF_Scores,AIFI_L3) %>% summarise(AverageScore=mean(value))

write.csv(df_combined_longformat,'NMF_Scores_SF4.csv')
