==> Figure5/01_NMF_CD4_T_cell_Projection/05_Get_Mean_expression_CD40LG_BRI.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)
library(reshape2)
source('/home//jupyter/BRI_Figures_Final_V1/helper_function/helper_function_IHA.r')

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA_meta_data-2024-05-09.csv")
meta_data_subset<-meta_data %>% filter(Covid_exclusion=='no',
                                       sample.visitName%in% c('Flu Year 1 Day 0'))

file_list<-paste0("/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA/BRI/Average_LogNormalized_Expression/Average_LogNormalized_Expression_of_Celltypes_by_Sample_AIFI_L3/", 
                  meta_data_subset$pbmc_sample_id, ".csv")

df_list<-read_pseudobulk_expression(file_list, mc_cores = 60)

# Subset CXCR5 from CM CD4 T

exp_long_format <- filter_genes_and_celltype(df_list, 'CD40LG', 'CM CD4 T cell', longformat = TRUE) 

exp_long_format<-left_join(exp_long_format,meta_data)

write.csv(exp_long_format,'CD40LG_CM_CD4T_BRI.csv')
