==> Figure5/01_NMF_CD4_T_cell_Projection/07_Get_Mean_expression_CD40_SF4.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)
library(reshape2)
source('/home//jupyter/BRI_Figures_Final_V2/helper_function/helper_function_IHA.r')

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V2/Dataset/SF4_meta_data-2024-05-05.csv")
meta_data$ageAtEnrollment[meta_data$ageAtEnrollment == '89+'] <- '90'
meta_data$ageAtEnrollment<-as.numeric(meta_data$ageAtEnrollment)
meta_data<-meta_data[!is.na(meta_data$ageAtEnrollment),]
meta_data$Ages<-meta_data$ageAtEnrollment
meta_data <- meta_data[!duplicated(meta_data$sample.sampleKitGuid, fromLast = TRUE), ]
meta_data$CMV<-meta_data$lab.revisionHistory.dataHistory.CMV.Ab.Screen.Result..CMVRES..newValue

file_list<-paste0("/home/jupyter/BRI_Figures_Final_V2/Dataset/scRNA/SF4/Average_LogNormalized_Expression/Average_LogNormalized_Expression_of_Celltypes_by_Sample_AIFI_L3/", 
                  meta_data$pbmc_sample_id, ".csv")

df_list<-read_pseudobulk_expression(file_list, mc_cores = 60)

# Subset CD40 from Core memory B

exp_long_format <- filter_genes_and_celltype(df_list, 'CD40', 'Core memory B cell', longformat = TRUE) 

exp_long_format<-left_join(exp_long_format,meta_data)

write.csv(exp_long_format,'CD40_CoreB_SF4.csv')
